upgrade v5.0.3

+++ client/app/component/admin/admin.component.ts
@@ -3,6 +3,7 @@ import { Component, OnInit } from '@angular/core';
 import { UserService } from '../../core/services/user.service';
 import { ToastComponent } from "../../shared/toast/toast.component";
 import { AuthService } from "../../core/services/auth.service";
+import { User } from '../../core/models/user.model';
 
 @Component({
   selector: 'app-admin',
@@ -11,7 +12,7 @@ import { AuthService } from "../../core/services/auth.service";
 })
 export class AdminComponent implements OnInit {
 
-  users = [];
+  users : User[] = [];
   isLoading = true;
 
   constructor(public auth: AuthService,
@@ -30,12 +31,14 @@ export class AdminComponent implements OnInit {
     );
   }
 
-  deleteUser(user) {
-    this.userService.deleteUser(user).subscribe(
-      data => this.toast.setMessage('user deleted successfully.', 'success'),
-      error => console.log(error),
-      () => this.getUsers()
-    );
+  deleteUser(user:User) {
+    if (window.confirm('Are you sure you want to delete ' + user.username + '?')) {
+           this.userService.deleteUser(user).subscribe(
+              data => this.toast.setMessage('user deleted successfully.', 'success'),
+              error => console.log(error),
+              () => this.getUsers()
+            );
+     }
   }
 
 }
